extract form logic into hook
